requirements.txt
Django
django-rest-framework
Pillow
psycopg2-binary
python-decouple
transliterate
django-rest-swagger
drf-yasg
django-filter
wheel